+++ coveralls/api.py
@@ -11,7 +11,6 @@ from .exception import CoverallsException
 from .git import git_info
 from .reporter import CoverallReporter
 
-
 log = logging.getLogger('coveralls.api')
 
 NUMBER_REGEX = re.compile(r'(\d+)$', re.IGNORECASE)
@@ -96,9 +95,8 @@ class Coveralls:
 
     @staticmethod
     def load_config_from_circle():
-        number = (
-            os.environ.get('CIRCLE_WORKFLOW_ID')
-            or os.environ.get('CIRCLE_BUILD_NUM')
+        number = os.environ.get('CIRCLE_WORKFLOW_ID') or os.environ.get(
+            'CIRCLE_BUILD_NUM',
         )
         pr = (os.environ.get('CI_PULL_REQUEST') or '').split('/')[-1] or None
         job = os.environ.get('CIRCLE_NODE_INDEX')
@@ -250,6 +248,7 @@ class Coveralls:
             with open(fname) as config:
                 try:
                     import yaml  # pylint: disable=import-outside-toplevel
+
                     self.config.update(yaml.safe_load(config))
                 except ImportError:
                     log.warning(
@@ -275,10 +274,12 @@ class Coveralls:
         return self.submit_report(json_string)
 
     def submit_report(self, json_string):
-        endpoint = f'{self._coveralls_host.rstrip("/")}/api/v1/jobs'
+        endpoint = f"{self._coveralls_host.rstrip('/')}/api/v1/jobs"
         verify = not bool(os.environ.get('COVERALLS_SKIP_SSL_VERIFY'))
         response = requests.post(
-            endpoint, files={'json_file': json_string}, verify=verify,
+            endpoint,
+            files={'json_file': json_string},
+            verify=verify,
         )
 
         if response.status_code == 422:
@@ -319,7 +320,7 @@ class Coveralls:
             # Github Actions only
             payload['repo_name'] = os.environ.get('GITHUB_REPOSITORY')
 
-        endpoint = f'{self._coveralls_host.rstrip("/")}/webhook'
+        endpoint = f"{self._coveralls_host.rstrip('/')}/webhook"
         verify = not bool(os.environ.get('COVERALLS_SKIP_SSL_VERIFY'))
         response = requests.post(endpoint, json=payload, verify=verify)
         try:
@@ -331,8 +332,9 @@ class Coveralls:
             ) from e
 
         if 'error' in response:
-            e = response['error']
-            raise CoverallsException(f'Parallel finish failed: {e}')
+            raise CoverallsException(
+                f"Parallel finish failed: {response['error']}",
+            )
 
         if 'done' not in response or not response['done']:
             raise CoverallsException('Parallel finish failed')
@@ -358,7 +360,8 @@ class Coveralls:
         log.debug('==\nReporting %s files\n==\n', len(data['source_files']))
         for source_file in data['source_files']:
             log.debug(
-                '%s - %d/%d', source_file['name'],
+                '%s - %d/%d',
+                source_file['name'],
                 sum(filter(None, source_file['coverage'])),
                 len(source_file['coverage']),
             )