+++ docs/conf.py
@@ -1,12 +1,11 @@
 from coveralls import __version__
 
-
 master_doc = 'index'
 source_suffix = '.rst'
 pygments_style = 'sphinx'
 
 templates_path = ['_templates']
-exclude_patterns = []
+exclude_patterns: list[str] = []
 
 todo_include_todos = True
 extensions = [