+++ nonunicode/malformed.py
@@ -1,4 +1,5 @@
 # -*- cоding: utf-8 -*-
 
+
 def hello():
     return 1