+++ tests/integration_test.py
@@ -6,7 +6,6 @@ import unittest
 
 from coveralls import Coveralls
 
-
 COVERAGE_CODE_STANZA = """
 import sys
 sys.path.append('{}')
@@ -49,10 +48,15 @@ class IntegrationTest(unittest.TestCase):
                     ),
                 )
 
-            subprocess.check_call([
-                sys.executable, '-m', 'coverage', 'run',
-                test_file,
-            ])
+            subprocess.check_call(
+                [
+                    sys.executable,
+                    '-m',
+                    'coverage',
+                    'run',
+                    test_file,
+                ],
+            )
 
             coverallz = Coveralls(repo_token='xxx')
             report = coverallz.create_data()
@@ -67,7 +71,8 @@ class IntegrationTest(unittest.TestCase):
                 (
                     f['coverage'] for f in report['source_files']
                     if f['name'] == foo
-                ), None,
+                ),
+                None,
             )
             assert sum(int(bool(x)) for x in lines) == hits